requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.8
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
Faker==4.0.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
waitress==1.4.3
whitenoise==5.0.1
